Added actions!

+++ src/com/sillypantscoder/pixeldungeon4/entities/TileEntity.java
@@ -4,7 +4,7 @@ import com.sillypantscoder.utils.JSON;
 import com.sillypantscoder.utils.Random;
 
 public abstract class TileEntity extends Entity {
-	public long id = Random.randomLong();
+	public int id = Random.randomInt();
 	public int x;
 	public int y;
 	public TileEntity(int time, int x, int y) {
